Add new useCurrentBranch parameter and refactor SPSEQ to use new representations (aside from pp)
