+++ src/main/java/de/upb/crypto/craco/sig/sps/eq/SPSEQSignature.java
@@ -4,8 +4,8 @@ import de.upb.crypto.craco.interfaces.signature.Signature;
 import de.upb.crypto.math.interfaces.structures.Group;
 import de.upb.crypto.math.interfaces.structures.GroupElement;
 import de.upb.crypto.math.serialization.Representation;
-import de.upb.crypto.math.serialization.annotations.AnnotatedRepresentationUtil;
-import de.upb.crypto.math.serialization.annotations.Represented;
+import de.upb.crypto.math.serialization.annotations.v2.ReprUtil;
+import de.upb.crypto.math.serialization.annotations.v2.Represented;
 
 import java.util.Objects;
 
@@ -20,32 +20,23 @@ public class SPSEQSignature implements Signature {
     /**
      * First group element of the signature in G_1.
      */
-    @Represented(structure = "groupG1", recoveryMethod = GroupElement.RECOVERY_METHOD)
+    @Represented(restorer = "G1")
     protected GroupElement group1ElementSigma1Z;
 
     /**
      * Second group element of the signature in G_1.
      */
-    @Represented(structure = "groupG1", recoveryMethod = GroupElement.RECOVERY_METHOD)
+    @Represented(restorer = "G1")
     protected GroupElement group1ElementSigma2Y;
 
     /**
      * Third group element of the signature in G_2.
      */
-    @Represented(structure = "groupG2", recoveryMethod = GroupElement.RECOVERY_METHOD)
+    @Represented(restorer = "G2")
     protected GroupElement group1ElementSigma3HatY;
 
-
-    // pointer fields used to store the structure for the representation process; in all other cases this should be null
-    protected Group groupG1 = null;
-    protected Group groupG2 = null;
-
     public SPSEQSignature(Representation repr, Group groupG1, Group groupG2) {
-        this.groupG1 = groupG1;
-        this.groupG2 = groupG2;
-        AnnotatedRepresentationUtil.restoreAnnotatedRepresentation(repr, this);
-        this.groupG1 = null;
-        this.groupG2 = null;
+        new ReprUtil(this).register(groupG1, "G1").register(groupG2, "G2"). deserialize(repr);
     }
 
     public SPSEQSignature(GroupElement group1ElementSigma1Z, GroupElement group1ElementSigma2Y, GroupElement group1ElementSigma3HatY) {
@@ -57,7 +48,7 @@ public class SPSEQSignature implements Signature {
 
     @Override
     public Representation getRepresentation() {
-        return AnnotatedRepresentationUtil.putAnnotatedRepresentation(this);
+        return ReprUtil.serialize(this);
     }
 
     public GroupElement getGroup1ElementSigma1Z() {
