+++ src/main/java/de/upb/crypto/craco/sig/sps/eq/SPSEQVerificationKey.java
@@ -4,12 +4,10 @@ import de.upb.crypto.craco.interfaces.signature.VerificationKey;
 import de.upb.crypto.math.interfaces.structures.Group;
 import de.upb.crypto.math.interfaces.structures.GroupElement;
 import de.upb.crypto.math.serialization.Representation;
-import de.upb.crypto.math.serialization.annotations.AnnotatedRepresentationUtil;
-import de.upb.crypto.math.serialization.annotations.Represented;
-import de.upb.crypto.math.serialization.annotations.RepresentedArray;
+import de.upb.crypto.math.serialization.annotations.v2.ReprUtil;
+import de.upb.crypto.math.serialization.annotations.v2.Represented;
 
 import java.util.Arrays;
-import java.util.Objects;
 
 /**
  * Class for the public (verification) key of the SPS-EQ signature scheme.
@@ -22,26 +20,20 @@ public class SPSEQVerificationKey implements VerificationKey {
     /**
      * \hat{X}_1, ..., \hat{X}_l \in G_2 in paper.
      */
-    @RepresentedArray(elementRestorer = @Represented(structure = "groupG2", recoveryMethod =
-            GroupElement.RECOVERY_METHOD))
+    @Represented(restorer = "[G2]")
     protected GroupElement[] group2ElementsHatXi;
-
-    // pointer field used to store the structure for the representation process; in all other cases this should be null
-    protected Group groupG2 = null;
 
     public SPSEQVerificationKey() {
         super();
     }
 
     public SPSEQVerificationKey(Group groupG2, Representation repr) {
-        this.groupG2 = groupG2;
-        AnnotatedRepresentationUtil.restoreAnnotatedRepresentation(repr, this);
-        this.groupG2 = null;
+        new ReprUtil(this).register(groupG2, "G2").deserialize(repr);
     }
 
     @Override
     public Representation getRepresentation() {
-        return AnnotatedRepresentationUtil.putAnnotatedRepresentation(this);
+        return ReprUtil.serialize(this);
     }
 
     public GroupElement[] getGroup2ElementsHatXi() {
